% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_rapport.R
\name{generer_rapport}
\alias{generer_rapport}
\title{Générer un rapport}
\usage{
generer_rapport(commune, departement, output)
}
\arguments{
\item{commune}{Nom de la commune}

\item{departement}{Nom du département}

\item{output}{Chemin du fichier de sortie}
}
\value{
Un fichier PDF ou HTML du rapport
}
\description{
Génère un rapport pour une commune et un département donnés
}
